--- Hangman.h
@@ -1,84 +1,0 @@
-#include <stdio.h>
-#include <stdbool.h>
-
-void Galgen() {
-	printf("_________\n");
-	printf("| /      \n");
-	printf("|/       \n");
-	printf("|        \n");
-	printf("|        \n");
-	printf("|        \n");
-	printf("|________\n");
-}
-void GalgenSeil() {
-	printf("_________\n");
-	printf("| /    | \n");
-	printf("|/       \n");
-	printf("|        \n");
-	printf("|        \n");
-	printf("|        \n");
-	printf("|________\n");
-}
-void GalgenKopf() {
-	printf("_________\n");
-	printf("| /    | \n");
-	printf("|/     O \n");
-	printf("|        \n");
-	printf("|        \n");
-	printf("|        \n");
-	printf("|________\n");
-}
-void GalgenBrust() {
-	printf("_________\n");
-	printf("| /    | \n");
-	printf("|/     O \n");
-	printf("|      | \n");
-	printf("|        \n");
-	printf("|        \n");
-	printf("|________\n");
-}
-void GalgenArmR() {
-	printf("_________\n");
-	printf("| /    | \n");
-	printf("|/     O \n");
-	printf("|      |\\\n");
-	printf("|        \n");
-	printf("|        \n");
-	printf("|________\n");
-}
-void GalgenArmRL() {
-	printf("_________\n");
-	printf("| /    | \n");
-	printf("|/     O \n");
-	printf("|     /|\\\n");
-	printf("|        \n");
-	printf("|        \n");
-	printf("|________\n");
-}
-void GalgenBeinR() {
-	printf("_________\n");
-	printf("| /    | \n");
-	printf("|/     O \n");
-	printf("|     /|\\\n");
-	printf("|       \\\n");
-	printf("|        \n");
-	printf("|________\n");
-}
-void GalgenBeinRL() {
-	printf("_________\n");
-	printf("| /    | \n");
-	printf("|/     O \n");
-	printf("|     /|\\\n");
-	printf("|     / \\\n");
-	printf("|        \n");
-	printf("|________\n");
-}
-void GalgenBeinRL() {
-	printf("######################################################\n");
-	printf("#   #   # #### ###  #       ##    ###  #### ##  #   #\n");
-	printf("#   #   # #    #  # #     ##  ##  #  # #    # # #   #\n");
-	printf("#   #   # #### ###  #    ##    ## ###  #### #  ##   #\n");
-	printf("#    # #  #    # #  #     ##  ##  # #  #    #  ##   #\n");
-	printf("#     #   #### #  # ####    ##    #  # #### #   #   #\n");
-	printf("######################################################\n");
-}
