--- Hangman.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.0.31919.166
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Hangman", "Hangman.vcxproj", "{F826DDCF-9B06-4E59-B30D-671C548A1F36}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{F826DDCF-9B06-4E59-B30D-671C548A1F36}.Debug|x64.ActiveCfg = Debug|x64
-		{F826DDCF-9B06-4E59-B30D-671C548A1F36}.Debug|x64.Build.0 = Debug|x64
-		{F826DDCF-9B06-4E59-B30D-671C548A1F36}.Debug|x86.ActiveCfg = Debug|Win32
-		{F826DDCF-9B06-4E59-B30D-671C548A1F36}.Debug|x86.Build.0 = Debug|Win32
-		{F826DDCF-9B06-4E59-B30D-671C548A1F36}.Release|x64.ActiveCfg = Release|x64
-		{F826DDCF-9B06-4E59-B30D-671C548A1F36}.Release|x64.Build.0 = Release|x64
-		{F826DDCF-9B06-4E59-B30D-671C548A1F36}.Release|x86.ActiveCfg = Release|Win32
-		{F826DDCF-9B06-4E59-B30D-671C548A1F36}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {ED990854-8469-412F-9516-D958FCCA8580}
-	EndGlobalSection
-EndGlobal
