+++ Hangman/Hangman.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.32002.261
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Hangman", "Hangman.vcxproj", "{E1D8C780-4140-4709-A7C1-A6A6AA45B07F}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{E1D8C780-4140-4709-A7C1-A6A6AA45B07F}.Debug|x64.ActiveCfg = Debug|x64
+		{E1D8C780-4140-4709-A7C1-A6A6AA45B07F}.Debug|x64.Build.0 = Debug|x64
+		{E1D8C780-4140-4709-A7C1-A6A6AA45B07F}.Debug|x86.ActiveCfg = Debug|Win32
+		{E1D8C780-4140-4709-A7C1-A6A6AA45B07F}.Debug|x86.Build.0 = Debug|Win32
+		{E1D8C780-4140-4709-A7C1-A6A6AA45B07F}.Release|x64.ActiveCfg = Release|x64
+		{E1D8C780-4140-4709-A7C1-A6A6AA45B07F}.Release|x64.Build.0 = Release|x64
+		{E1D8C780-4140-4709-A7C1-A6A6AA45B07F}.Release|x86.ActiveCfg = Release|Win32
+		{E1D8C780-4140-4709-A7C1-A6A6AA45B07F}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {E827D534-3D01-4EFB-B6C2-E7512D129D3A}
+	EndGlobalSection
+EndGlobal
