+++ Hangman/Quelle.c
@@ -1,0 +1,6 @@
+#include <stdio.h>
+
+int main() {
+
+	printf("Hello");
+}
